grus: Use boot_devices parameter for device fstab
This is a thing since introduction of dynamic partitioning with Android 10.
It's used to allow init to enable creation of /dev/block/by-name symlinks.

This also benefits devices not using dynamic partitions as well since fstab
changes for first stage init of required file systems are no longer reqiring
full soc path to be defined. Win-win.

Link: https://source.android.com/devices/tech/ota/dynamic_partitions/implement#kernel-command-line-changes
Signed-off-by: Albert I <kras@raphielgang.org>